tracing: add flags to reject policies with signal and override actions
This patch allows disabling of the 2 action types that allow
modification of syscall behaviour.
This is desirable in certain scenarios to prevent privilage escalation
via loading policies via the gRPC API.

This patch implements this by adding 2 flags to the tetragon binary
  --disable-signal-actions which causes tetragon to reject any policy
  that specifies `Signal` or `Sigkill` actions
  --disable-override-actions which causes tetragon to reject any policy
  that specifies `Override` actions

Rejection was preferred over silent removal of the disabled actions from
the policy as that will have less unexpected behaviour.

Signed-off-by: Ben Segall <bentekkie@google.com>